feat(#37 - Analytics) - Add page tracking and contact link tracing PT 1
Adds in Google Analytics page tracking and click tracking to the app.
Production test will be performed by Steve Caspers to confirm it is
performing as expected, since I don't have a local setup environment for this.

Bonus: Made login background Murphy Auto red to spice it up a bit.