fix(#37 - Analytics) - Manually insert GA tag instead of async document.write() thru library
The existing library we are using is no longer compatible with Angular CLI when building.
Because of this, the call which writes the script tag to index.html no longer works.
The only reason the existing deployed apps work is that they haven't been updated since
the Angular 8 transition.

I will need to make a change to both projects going forward to make it handle this so
that when I am developing, I don't pollute production analytics.