feat(#23 - Vehicle History/Inspections/Billing) - Select vehicle in history via dropdown
Added dropdown for selecting vehicle(s) in history. The initial state from the URL
sets the dropdown selection and chooses which vehicles are in view.
Following that, the select box allows further changes in case the vehicle the user
selected upon initial navigation was incorrect.

The vehicle list in the dropdown should be sorted by year, make, model, and license.
Vehicles which do not yet have history (because they have no VIN) are omitted,